Move onion requests to typescript
